+++ packages/api/models/granules.js
@@ -7,6 +7,7 @@ const uniqBy = require('lodash.uniqby');
 
 const aws = require('@cumulus/ingest/aws');
 const commonAws = require('@cumulus/common/aws');
+const StepFunctions = require('@cumulus/common/StepFunctions');
 const cmrjs = require('@cumulus/cmrjs');
 const { CMR, reconcileCMRMetadata } = require('@cumulus/cmrjs');
 const log = require('@cumulus/common/log');
@@ -16,7 +17,6 @@ const {
   moveGranuleFiles
 } = require('@cumulus/ingest/granule');
 const { constructCollectionId } = require('@cumulus/common');
-const { describeExecution } = require('@cumulus/common/step-functions');
 
 const Manager = require('./base');
 
@@ -95,7 +95,7 @@ class Granule extends Manager {
   async reingest(granule) {
     const executionArn = path.basename(granule.execution);
 
-    const executionDescription = await describeExecution(executionArn);
+    const executionDescription = await StepFunctions.describeExecution({ executionArn });
     const originalMessage = JSON.parse(executionDescription.input);
 
     const { name, version } = deconstructCollectionId(granule.collectionId);